+++ sudachipy/plugin/path_rewrite/numericparser.py
